Extract CommonMessages to persistence-model
