+++ persistence/src/main/scala/hmda/persistence/LocalEventPublisher.scala
@@ -1,6 +1,6 @@
 package hmda.persistence
 import akka.actor.ActorSystem
-import hmda.persistence.CommonMessages.Event
+import hmda.persistence.messages.CommonMessages.Event
 
 trait LocalEventPublisher extends EventPublisher {
