+++ persistence/src/test/scala/hmda/persistence/institutions/FilingPersistenceSpec.scala
@@ -3,7 +3,7 @@ package hmda.persistence.institutions
 import akka.testkit.{ EventFilter, TestProbe }
 import hmda.actor.test.ActorSpec
 import hmda.model.fi._
-import hmda.persistence.CommonMessages.GetState
+import hmda.persistence.messages.CommonMessages.GetState
 import hmda.persistence.demo.DemoData
 import hmda.persistence.institutions.FilingPersistence._
